requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
cloudpickle==3.1.0
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
gymnasium==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.4
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-core==0.3.24
langchain-text-splitters==0.3.2
langsmith==0.2.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
probableparsing==0.0.1
propcache==0.2.1
pyap==0.3.1
pydantic==2.9.2
pydantic_core==2.23.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
# together  ← replaced by local Qwen model (no API needed)
# openai    ← no longer required (local inference)
torch==2.5.1
tqdm==4.67.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
usaddress==0.5.11
virtualenv==20.28.0
wget==3.2
yarl==1.18.3
langchain_experimental
vowpal_wabbit_next

# HuggingFace — local VLM + LLM inference
transformers>=4.40.0
accelerate>=0.27.0
huggingface-hub>=0.22.0
sentencepiece
einops

# Jupyter/IPython
ipython
jupyter

# PDF processing
pdfplumber
PyPDF2
PyMuPDF
pdf2image          # PDF → PIL images for VLM pathway

# Document processing
python-docx
docx2txt

# Image processing
Pillow
opencv-python
paddleocr
paddlepaddle
